Switch to using a command line script
