+++ package.json
@@ -14,11 +14,11 @@
     "eslint": "eslint .",
     "stylelint": "stylelint --syntax scss **/*.scss",
     "lint": "npm run eslint && npm run stylelint",
-    "servertest": "ADDONS_FRONTEND_BUILD_ALL=1 npm run build && better-npm-run servertest && better-npm-run servertest:amo && better-npm-run servertest:disco && better-npm-run servertest:search",
+    "servertest": "bin/config-check.js && ADDONS_FRONTEND_BUILD_ALL=1 npm run build && better-npm-run servertest && better-npm-run servertest:amo && better-npm-run servertest:disco && better-npm-run servertest:search",
     "start": "npm run version-check && NODE_PATH='./:./src' node bin/server.js",
-    "test": "better-npm-run test",
-    "unittest": "better-npm-run unittest",
-    "unittest:dev": "better-npm-run unittest:dev",
+    "test": "bin/config-check.js && better-npm-run test",
+    "unittest": "bin/config-check.js && better-npm-run unittest",
+    "unittest:dev": "bin/config-check.js && better-npm-run unittest:dev",
     "version-check": "node bin/version-check.js",
     "webpack-dev-server": "better-npm-run webpack-dev-server"
   },
